requirements.txt
textblob==0.18.0.post0
gradio
requests
scikit-learn
python-dotenv
matplotlib